Add alexandria.search, an interface for searching the database
search.perform_simple_search searches the document index by splitting
the search string based on the same separators as the index's map function.

It then retrieves each set of results for each of the search keys, performs
an intersection of all sets, and then returns the results as a set of tuples.

Each tuple is of the form (host_document_key, file_hash) where
host_document_key is the _id of the host document containing that file,
and file_hash is the MD5 of the full path (minus hostname) to the file, which
is also the key to that file in the document's `file' associative array.