Add `pushviews' option to manage_alexandria.py
This option will push all design documents specified in action_pushviews to
CouchDB, including those that have not changed. In the future, we should do
some sort of check (maybe MD5?) to avoid pushing views that will take a long
time to completely rebuild (which will happen after this option is run).